Kotlin: Convert NetworkTaskManager to Kotlin

+++ mycuration/src/main/java/com/phicdy/mycuration/presenter/FeedSearchPresenter.java
@@ -82,7 +82,7 @@ public class FeedSearchPresenter implements Presenter {
         }
     }
 
-    void onFinishAddFeed(@NonNull String url, int reason) {
+    void onFinishAddFeed(@NonNull String url, long reason) {
         Feed newFeed = adapter.getFeedByUrl(url);
         if (reason == RssParseResult.NOT_FAILED && newFeed != null) {
             unreadManager.addFeed(newFeed);
